Remove signup feature under user feature
